requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2021.10.8
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.29.0
colorama==0.4.4
Flask==2.1.2
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.4
mysql-connector-python==8.0.28
nextcord==2.3.3
Pillow==9.4.0
protobuf==4.21.1
python-dotenv==0.20.0
six==1.16.0
typing==3.7.4.3
typing_extensions==4.4.0
urllib3==1.26.9
Werkzeug==2.1.2
-e git+https://mygit.wmg.warwick.ac.uk/u2190045/wm393-sdlc-assignment-2.git@5bb366485c59dd3aaebf05b999a3a6d4e6cb19ca#egg=wmgsis
yarl==1.8.2
